% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/and.R
\name{and.gddMatch}
\alias{and.gddMatch}
\title{Combine gddMatch Queries}
\usage{
\method{and}{gddMatch}(x, ..., rows = FALSE)
}
\arguments{
\item{x}{A \code{gddMatch} object returned by the \code{gddMatch} function.}

\item{...}{More \code{gddMatch} elements.}

\item{rows}{A boolean (TRUE/FALSE), should the actual text rows be added to the resulting object?}
}
\description{
Combine results from \code{gddMatch} into a single object.
}
