% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/and.R
\name{and}
\alias{and}
\title{Combine gddMatch Queries}
\usage{
and(x, ..., rows = FALSE)
}
\arguments{
\item{x}{A \code{gddMatch} object returned by the \code{gddMatch} function.}

\item{...}{More \code{gddMatch} elements.}

\item{rows}{A boolean (TRUE/FALSE), should the actual text rows be added to the resulting object?}
}
\description{
Working towards combining records in GDD.
}
