% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse}
\alias{browse}
\title{Summarize publications}
\usage{
browse(x, corpus = NULL, pubs = NULL, words = NULL)
}
\arguments{
\item{x}{One of, a vector of character \code{gddid}s from GeoDeepDive, or a \code{data.frame} with columns \code{gddid} and \code{words}.}

\item{corpus}{The full GDD corpus of which \code{x} is expected to be a subset.}

\item{pubs}{The bibliographic information for the GDD dataset.}

\item{words}{A vector of sentences, similar to the GDD corpus \code{words} column.}
}
\description{
A quick view method for looking at publication data from the GeoDeepDive publications.
}
\examples{
\dontrun{
subset <- sample(nrow(nlp), 100)
browse(x = nlp$`_gddid`[subset],
       pubs = publications,
       words = nlp$word[subset])
}
}
