% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_words.R
\name{clean_words}
\alias{clean_words}
\title{Clean words}
\usage{
clean_words(x)
}
\arguments{
\item{x}{A vector from the NLP \code{data.frame}}
}
\value{
A \code{character} vector of length equivalent to \code{x}.
}
\description{
A function to render the word vector to the OCRed sentence.
}
