% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dd.R
\name{load_dd}
\alias{load_dd}
\title{Load deepdive dump to SQL database}
\usage{
load_dd(con, bib, sent, bib_name = "publications", sent_name = "sentences",
  clean = TRUE)
}
\arguments{
\item{con}{An object returned from \code{dbConnect()}}

\item{bib}{A bibjson file exported from GeoDeepDive.}

\item{sent}{A set of parsed sentence files from the NLP output.}

\item{bib_name}{The name for the bibliography table in the database.}

\item{sent_name}{The name for the sentence table in the database.}
}
\description{
Takes a raw SQL file and loads it into a Postgres database.
}
\examples{
library(geodiveR)
library(RPostgreSQL)
  con <- dbConnect(drv = "PostgreSQL",
                   user = "postgres",
                   password = "postgres",
                   host = "localhost",
                   port = "5432",
                   dbname = "deepdive")

  # Using the package's data:
  data(nlp)
  data(publications)

  # Or alternately a string to the file path:
  # nlp <- "./data/sentences_nlp352"
  # publications <- "./data/bibjson"

  load_dd(con, bib = publications, sent = nlp)

}
