% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryGdd.R
\name{summaryGdd}
\alias{summaryGdd}
\title{Return database level GDD Info}
\usage{
summaryGdd(con)
}
\arguments{
\item{con}{GDD database connection.}
}
\value{
A data.frame
}
\details{
Return information about the Postgres GDD database.
}
\examples{
#' library(geodiveR)
library(RPostgreSQL)
  con <- dbConnect(drv = "PostgreSQL",
                   user = "postgres",
                   password = "postgres",
                   host = "localhost",
                   port = "5432",
                   dbname = "deepdive")
 summary(con)

}
