% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse}
\alias{browse}
\title{Summarize publications}
\usage{
browse(x, corpus = NULL, pubs = NULL, words = NULL)
}
\arguments{
\item{x}{A vector of numeric \code{gddid}s from GeoDeepDive, a \code{boolean} vector where \code{length(x) == corpus}, or a \code{data.frame} with columns \code{gddid} and \code{words}.}

\item{corpus}{The full GDD corpus of which \code{x} is expected to be a subset.}

\item{pubs}{The bibliographic information for the GDD dataset.}

\item{words}{A vector of sentences, similar to the GDD corpus \code{words} column.}
}
\description{
A quick view method for looking at publication data from the GeoDeepDive publications.
}
\examples{
library(geodiveR)
library(RPostgreSQL)
  con <- dbConnect(drv = "PostgreSQL",
                   user = "postgres",
                   password = "postgres",
                   host = "localhost",
                   port = "5432",
                   dbname = "deepdive")

# Return only the publication information, using gddids as the key:

# Using a numeric or boolean index of sentences:

 july_sent <- gddMatch(con,
                       table = 'sentences',
                       col = 'words',
                       pattern = '((July)|(August))',
                       boolean = TRUE,
                       name = 'SummerQuery')

browse(july_sent,
       pubs = TRUE)

}
