% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gddMatch.R
\name{gddMatch}
\alias{gddMatch}
\title{Build matches using basic pattern matching.}
\usage{
gddMatch(con, table, col, pattern, rows = FALSE, name)
}
\arguments{
\item{con}{A database connection to the GDD database.}

\item{table}{The table of interest.}

\item{col}{The table column.}

\item{pattern}{A Postgres-compliant (and case-sensitive) regex pattern. \url{https://www.postgresql.org/docs/9.3/static/functions-matching.html}.}

\item{rows}{Do you want to return the explicit rows (default is FALSE)?}

\item{name}{The name used to identify the query.}
}
\description{
A wrapper for the \code{DBI::dbGetQuery()} that provides support for basic pattern matching.
}
\examples{
library(geodiveR)
library(DBI)
  con <- dbConnect(drv = "PostgreSQL",
                   user = "postgres",
                   password = "postgres",
                   host = "localhost",
                   port = "5432",
                   dbname = "deepdive")

 july_sent <- gddMatch(con = con,
                       table = 'sentences',
                       col = 'words',
                       pattern = 'July',
                       name = "JulyQuery")

}
